fix bug

+++ environment.py
@@ -379,9 +379,9 @@ class Environment:
 
         '''
         obs = np.zeros((self.num_agents, 2, 2*self.obs_radius+1, 2*self.obs_radius+1), dtype=np.bool)
-        pos = np.zeros((self.num_agents, 2), dtype=np.uint8)
+        pos = np.zeros((self.num_agents, 2), dtype=np.int16)
 
-        pos = self.goals_pos-self.agents_pos
+        pos = self.goals_pos.astype(np.int16)-self.agents_pos.astype(np.int16)
 
         # 0 represents obstacle to match 0 padding in CNN 
         obstacle_map = np.pad(self.map==0, self.obs_radius, 'constant', constant_values=0)
